+++ android/src/main/java/com/reactnativestripesdk/StripeSdkPackage.kt
@@ -4,6 +4,7 @@ import com.facebook.react.ReactPackage
 import com.facebook.react.bridge.NativeModule
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.uimanager.ViewManager
+import com.reactnativestripesdk.pushprovisioning.AddToWalletButtonManager
 
 class StripeSdkPackage : ReactPackage {
     override fun createNativeModules(reactContext: ReactApplicationContext): List<NativeModule> {
